+++ src/index.js
@@ -5,8 +5,6 @@ import { XAiClient } from './xai.js';
 
 export class Client {
   constructor(options = {}) {
-    const { platform } = options;
-
     if (!options.platform) {
       throw new Error('No platform specified.');
     } else if (!options.templates) {
@@ -14,17 +12,61 @@ export class Client {
     } else if (!options.apiKey) {
       throw new Error('No API key specified.');
     }
+    return getClientForPlatform(options);
+  }
+}
 
-    if (platform === 'openai' || platform === 'gpt') {
-      return new OpenAiClient(options);
-    } else if (platform === 'google' || platform === 'gemini') {
-      return new GoogleClient(options);
-    } else if (platform === 'anthropic' || platform === 'claude') {
-      return new AnthropicClient(options);
-    } else if (platform === 'xai' || platform === 'grok') {
-      return new XAiClient(options);
-    } else if (platform) {
-      throw new Error(`Unknown platform "${platform}".`);
+export class MultiClient {
+  constructor(options) {
+    const { platforms } = options;
+
+    this.clients = {};
+
+    for (let platform of platforms) {
+      const { name, apiKey } = platform;
+      const client = getClientForPlatform({
+        ...options,
+        platform: name,
+        apiKey,
+      });
+      this.clients[name] = client;
+      this.clients[undefined] ||= client;
     }
+  }
+
+  prompt(options) {
+    return this.getClient(options).prompt(options);
+  }
+
+  stream(options) {
+    return this.getClient(options).stream(options);
+  }
+
+  buildTemplate(options) {
+    return this.getClient(options).buildTemplate(options);
+  }
+
+  getClient(options) {
+    const { platform } = options;
+    const client = this.clients[platform];
+    if (!client) {
+      throw new Error(`Platform "${platform}" not found.`);
+    }
+    return client;
+  }
+}
+
+function getClientForPlatform(options) {
+  const { platform } = options;
+  if (platform === 'openai' || platform === 'gpt') {
+    return new OpenAiClient(options);
+  } else if (platform === 'google' || platform === 'gemini') {
+    return new GoogleClient(options);
+  } else if (platform === 'anthropic' || platform === 'claude') {
+    return new AnthropicClient(options);
+  } else if (platform === 'xai' || platform === 'grok') {
+    return new XAiClient(options);
+  } else if (platform) {
+    throw new Error(`Unknown platform "${platform}".`);
   }
 }