+++ test/entry.test.js
@@ -3,12 +3,12 @@ import path from 'path';
 import { setResponse as setGptResponse } from 'openai';
 import { setResponse as setClaudeResponse } from '@anthropic-ai/sdk';
 
-import { Client } from '../src/index';
+import { Client, MultiClient } from '../src/index';
 
 import gpt from './responses/openai/text.json';
 import claude from './responses/anthropic/text.json';
 
-describe('entry', () => {
+describe('Client', () => {
   it('should error on bad setup', async () => {
     expect(() => {
       new Client();
@@ -78,6 +78,56 @@ describe('entry', () => {
       { name: 'NVIDIA', symbol: 'NVDA' },
       { name: 'Alphabet (Google)', symbol: 'GOOGL' },
       { name: 'Amazon', symbol: 'AMZN' },
+    ]);
+  });
+});
+
+describe('MultiClient', () => {
+  it('should set multiple keys at once', async () => {
+    setGptResponse(gpt);
+    setClaudeResponse(claude);
+
+    const client = new MultiClient({
+      templates: path.join(__dirname, './templates'),
+      platforms: [
+        {
+          name: 'claude',
+          apiKey: 'claude-key',
+        },
+        {
+          name: 'openai',
+          apiKey: 'openai-key',
+        },
+      ],
+    });
+
+    let result;
+
+    result = await client.prompt({
+      file: 'stocks',
+      output: 'json',
+    });
+
+    expect(result).toEqual([
+      { name: 'Microsoft', symbol: 'MSFT' },
+      { name: 'Apple', symbol: 'AAPL' },
+      { name: 'NVIDIA', symbol: 'NVDA' },
+      { name: 'Alphabet (Google)', symbol: 'GOOGL' },
+      { name: 'Amazon', symbol: 'AMZN' },
+    ]);
+
+    result = await client.prompt({
+      file: 'stocks',
+      output: 'json',
+      platform: 'openai',
+    });
+
+    expect(result).toEqual([
+      { name: 'Apple Inc.', symbol: 'AAPL' },
+      { name: 'Microsoft Corporation', symbol: 'MSFT' },
+      { name: 'Amazon.com Inc.', symbol: 'AMZN' },
+      { name: 'NVIDIA Corporation', symbol: 'NVDA' },
+      { name: 'Alphabet Inc. (Class A)', symbol: 'GOOGL' },
     ]);
   });
 });